01version/backend_fastapi/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
motor
requests
python-multipart